refator: extract the segment with weight module
